+++ src/steps/account/index.ts
@@ -5,14 +5,20 @@ import {
 import { IntegrationConfig } from '../../config';
 import { Entities, SetDataKeys, StepIds } from '../../constants';
 import { createServiceEntity } from '../../converters';
+import { APIClient } from '../../snyk/client';
 
 async function fetchAccount(
   context: IntegrationStepExecutionContext<IntegrationConfig>,
 ) {
-  const { instance, jobState } = context;
+  const { instance, jobState, logger } = context;
+
+  const apiClient = new APIClient(logger, instance.config);
+  const currentOrgName = await apiClient.getCurrentOrgName();
+
   const serviceEntity = await jobState.addEntity(
-    createServiceEntity(instance.config.snykOrgId),
+    createServiceEntity(instance.config.snykOrgId, currentOrgName),
   );
+
   await jobState.setData(SetDataKeys.ACCOUNT_ENTITY, serviceEntity);
 }